function -> hook + some refactor around it
